+++ modules/openapi-generator/src/main/resources/swift-alt/Old.swift
@@ -1,0 +1,1361 @@
+//
+// PetAPI.swift
+//
+// Generated by openapi-generator
+// https://openapi-generator.tech
+//
+
+import Foundation
+import Combine
+// import Runtime
+
+
+open class PetAPI {
+    private let transport: Transport
+
+    public init(_ transport: Transport) {
+        self.transport = transport
+    }
+    /**
+        Add a new pet to the store
+    - POST /pet
+        - OAuth:
+        - type: oauth2
+        - name: petstore_auth
+        - parameter pet: (body) Pet object that needs to be added to the store
+    - returns: Future<Pet, Never>
+        */
+    open func addPet(pet: Pet) -> AnyPublisher<Result<Pet, Error>, Never> {
+        var request = URLRequest(url: transport.baseURL)
+        return transport.send(request: request, authType: .none)
+            .eraseToAnyPublisher()
+            
+    }
+    /**
+        Deletes a pet
+    - DELETE /pet/{petId}
+        - OAuth:
+        - type: oauth2
+        - name: petstore_auth
+        - parameter petId: (path) Pet id to delete
+        - parameter apiKey: (header)  (optional)
+    - returns: Future<Void, Never>
+        */
+    open func deletePet(petId: Int64, apiKey: String? = nil) -> Future<Void, Never> {
+    }
+    /**
+        Finds Pets by status
+    - GET /pet/findByStatus
+        - Multiple status values can be provided with comma separated strings
+        - OAuth:
+        - type: oauth2
+        - name: petstore_auth
+        - parameter status: () Status values that need to be considered for filter
+    - returns: Future<[Pet], Never>
+        */
+    open func findPetsByStatus(status: [String]) -> Future<[Pet], Never> {
+    }
+    /**
+        Finds Pets by tags
+    - GET /pet/findByTags
+        - Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.
+        - OAuth:
+        - type: oauth2
+        - name: petstore_auth
+        - parameter tags: () Tags to filter by
+    - returns: Future<[Pet], Never>
+        */
+    open func findPetsByTags(tags: [String]) -> Future<[Pet], Never> {
+    }
+    /**
+        Find pet by ID
+    - GET /pet/{petId}
+        - Returns a single pet
+        - API Key:
+        - type: apiKey api_key
+        - name: api_key
+        - parameter petId: (path) ID of pet to return
+    - returns: Future<Pet, Never>
+        */
+    open func getPetById(petId: Int64) -> Future<Pet, Never> {
+    }
+    /**
+        Update an existing pet
+    - PUT /pet
+        - OAuth:
+        - type: oauth2
+        - name: petstore_auth
+        - parameter pet: (body) Pet object that needs to be added to the store
+    - returns: Future<Pet, Never>
+        */
+    open func updatePet(pet: Pet) -> Future<Pet, Never> {
+    }
+    /**
+        Updates a pet in the store with form data
+    - POST /pet/{petId}
+        - OAuth:
+        - type: oauth2
+        - name: petstore_auth
+        - parameter petId: (path) ID of pet that needs to be updated
+        - parameter name: (form) Updated name of the pet (optional)
+        - parameter status: (form) Updated status of the pet (optional)
+    - returns: Future<Void, Never>
+        */
+    open func updatePetWithForm(petId: Int64, name: String? = nil, status: String? = nil) -> Future<Void, Never> {
+    }
+    /**
+        uploads an image
+    - POST /pet/{petId}/uploadImage
+        - OAuth:
+        - type: oauth2
+        - name: petstore_auth
+        - parameter petId: (path) ID of pet to update
+        - parameter additionalMetadata: (form) Additional data to pass to server (optional)
+        - parameter file: (form) file to upload (optional)
+    - returns: Future<ApiResponse, Never>
+        */
+    open func uploadFile(petId: Int64, additionalMetadata: String? = nil, file: Data? = nil) -> Future<ApiResponse, Never> {
+    }
+}
+////
+
+/**
+-
+    - API Key:
+    - type: apiKey api_key
+    - name: api_key
+    - OAuth:
+    - type: oauth2
+    - name: petstore_auth
+- returns: RequestBuilder<Void>
+    */
+open class func WithRequestBuilder() -> RequestBuilder<Void> {
+let localVariablePath = ""
+    let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+            let localVariableParameters: [String: Any]? = nil
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+    let localVariableNillableHeaders: [String: Any?] = [
+        :
+    ]
+
+    let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+    let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+    return localVariableRequestBuilder.init(method: "", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+import Foundation
+#if canImport(AnyCodable)
+import AnyCodable
+#endif
+
+open class PetAPI {
+
+    /**
+     Add a new pet to the store
+     
+     - parameter pet: (body) Pet object that needs to be added to the store
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func addPet(pet: Pet, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Pet?, _ error: Error?) -> Void)) {
+        addPetWithRequestBuilder(pet: pet).execute(apiResponseQueue) { result in
+            switch result {
+            case let .success(response):
+                completion(response.body, nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Add a new pet to the store
+     - POST /pet
+     - OAuth:
+       - type: oauth2
+       - name: petstore_auth
+     - parameter pet: (body) Pet object that needs to be added to the store
+     - returns: RequestBuilder<Pet>
+     */
+    open class func addPetWithRequestBuilder(pet: Pet) -> RequestBuilder<Pet> {
+        let localVariablePath = "/pet"
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: pet)
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Pet>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
+
+        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Deletes a pet
+     
+     - parameter petId: (path) Pet id to delete
+     - parameter apiKey: (header)  (optional)
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func deletePet(petId: Int64, apiKey: String? = nil, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
+        deletePetWithRequestBuilder(petId: petId, apiKey: apiKey).execute(apiResponseQueue) { result in
+            switch result {
+            case .success:
+                completion((), nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Deletes a pet
+     - DELETE /pet/{petId}
+     - OAuth:
+       - type: oauth2
+       - name: petstore_auth
+     - parameter petId: (path) Pet id to delete
+     - parameter apiKey: (header)  (optional)
+     - returns: RequestBuilder<Void>
+     */
+    open class func deletePetWithRequestBuilder(petId: Int64, apiKey: String? = nil) -> RequestBuilder<Void> {
+        var localVariablePath = "/pet/{petId}"
+        let petIdPreEscape = "\(APIHelper.mapValueToPathItem(petId))"
+        let petIdPostEscape = petIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
+        localVariablePath = localVariablePath.replacingOccurrences(of: "{petId}", with: petIdPostEscape, options: .literal, range: nil)
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters: [String: Any]? = nil
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            "api_key": apiKey?.encodeToJSON(),
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+        return localVariableRequestBuilder.init(method: "DELETE", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     * enum for parameter status
+     */
+    public enum Status_findPetsByStatus: String, CaseIterable {
+        case available = "available"
+        case pending = "pending"
+        case sold = "sold"
+    }
+
+    /**
+     Finds Pets by status
+     
+     - parameter status: (query) Status values that need to be considered for filter
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func findPetsByStatus(status: [String], apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: [Pet]?, _ error: Error?) -> Void)) {
+        findPetsByStatusWithRequestBuilder(status: status).execute(apiResponseQueue) { result in
+            switch result {
+            case let .success(response):
+                completion(response.body, nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Finds Pets by status
+     - GET /pet/findByStatus
+     - Multiple status values can be provided with comma separated strings
+     - OAuth:
+       - type: oauth2
+       - name: petstore_auth
+     - parameter status: (query) Status values that need to be considered for filter
+     - returns: RequestBuilder<[Pet]>
+     */
+    open class func findPetsByStatusWithRequestBuilder(status: [String]) -> RequestBuilder<[Pet]> {
+        let localVariablePath = "/pet/findByStatus"
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters: [String: Any]? = nil
+
+        var localVariableUrlComponents = URLComponents(string: localVariableURLString)
+        localVariableUrlComponents?.queryItems = APIHelper.mapValuesToQueryItems([
+            "status": status.encodeToJSON(),
+        ])
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<[Pet]>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
+
+        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Finds Pets by tags
+     
+     - parameter tags: (query) Tags to filter by
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    @available(*, deprecated, message: "This operation is deprecated.")
+    open class func findPetsByTags(tags: [String], apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: [Pet]?, _ error: Error?) -> Void)) {
+        findPetsByTagsWithRequestBuilder(tags: tags).execute(apiResponseQueue) { result in
+            switch result {
+            case let .success(response):
+                completion(response.body, nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Finds Pets by tags
+     - GET /pet/findByTags
+     - Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.
+     - OAuth:
+       - type: oauth2
+       - name: petstore_auth
+     - parameter tags: (query) Tags to filter by
+     - returns: RequestBuilder<[Pet]>
+     */
+    @available(*, deprecated, message: "This operation is deprecated.")
+    open class func findPetsByTagsWithRequestBuilder(tags: [String]) -> RequestBuilder<[Pet]> {
+        let localVariablePath = "/pet/findByTags"
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters: [String: Any]? = nil
+
+        var localVariableUrlComponents = URLComponents(string: localVariableURLString)
+        localVariableUrlComponents?.queryItems = APIHelper.mapValuesToQueryItems([
+            "tags": tags.encodeToJSON(),
+        ])
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<[Pet]>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
+
+        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Find pet by ID
+     
+     - parameter petId: (path) ID of pet to return
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func getPetById(petId: Int64, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Pet?, _ error: Error?) -> Void)) {
+        getPetByIdWithRequestBuilder(petId: petId).execute(apiResponseQueue) { result in
+            switch result {
+            case let .success(response):
+                completion(response.body, nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Find pet by ID
+     - GET /pet/{petId}
+     - Returns a single pet
+     - API Key:
+       - type: apiKey api_key
+       - name: api_key
+     - parameter petId: (path) ID of pet to return
+     - returns: RequestBuilder<Pet>
+     */
+    open class func getPetByIdWithRequestBuilder(petId: Int64) -> RequestBuilder<Pet> {
+        var localVariablePath = "/pet/{petId}"
+        let petIdPreEscape = "\(APIHelper.mapValueToPathItem(petId))"
+        let petIdPostEscape = petIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
+        localVariablePath = localVariablePath.replacingOccurrences(of: "{petId}", with: petIdPostEscape, options: .literal, range: nil)
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters: [String: Any]? = nil
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Pet>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
+
+        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Update an existing pet
+     
+     - parameter pet: (body) Pet object that needs to be added to the store
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func updatePet(pet: Pet, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Pet?, _ error: Error?) -> Void)) {
+        updatePetWithRequestBuilder(pet: pet).execute(apiResponseQueue) { result in
+            switch result {
+            case let .success(response):
+                completion(response.body, nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Update an existing pet
+     - PUT /pet
+     - OAuth:
+       - type: oauth2
+       - name: petstore_auth
+     - parameter pet: (body) Pet object that needs to be added to the store
+     - returns: RequestBuilder<Pet>
+     */
+    open class func updatePetWithRequestBuilder(pet: Pet) -> RequestBuilder<Pet> {
+        let localVariablePath = "/pet"
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: pet)
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Pet>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
+
+        return localVariableRequestBuilder.init(method: "PUT", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Updates a pet in the store with form data
+     
+     - parameter petId: (path) ID of pet that needs to be updated
+     - parameter name: (form) Updated name of the pet (optional)
+     - parameter status: (form) Updated status of the pet (optional)
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func updatePetWithForm(petId: Int64, name: String? = nil, status: String? = nil, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
+        updatePetWithFormWithRequestBuilder(petId: petId, name: name, status: status).execute(apiResponseQueue) { result in
+            switch result {
+            case .success:
+                completion((), nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Updates a pet in the store with form data
+     - POST /pet/{petId}
+     - OAuth:
+       - type: oauth2
+       - name: petstore_auth
+     - parameter petId: (path) ID of pet that needs to be updated
+     - parameter name: (form) Updated name of the pet (optional)
+     - parameter status: (form) Updated status of the pet (optional)
+     - returns: RequestBuilder<Void>
+     */
+    open class func updatePetWithFormWithRequestBuilder(petId: Int64, name: String? = nil, status: String? = nil) -> RequestBuilder<Void> {
+        var localVariablePath = "/pet/{petId}"
+        let petIdPreEscape = "\(APIHelper.mapValueToPathItem(petId))"
+        let petIdPostEscape = petIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
+        localVariablePath = localVariablePath.replacingOccurrences(of: "{petId}", with: petIdPostEscape, options: .literal, range: nil)
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableFormParams: [String: Any?] = [
+            "name": name?.encodeToJSON(),
+            "status": status?.encodeToJSON(),
+        ]
+
+        let localVariableNonNullParameters = APIHelper.rejectNil(localVariableFormParams)
+        let localVariableParameters = APIHelper.convertBoolToString(localVariableNonNullParameters)
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            "Content-Type": "application/x-www-form-urlencoded",
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     uploads an image
+     
+     - parameter petId: (path) ID of pet to update
+     - parameter additionalMetadata: (form) Additional data to pass to server (optional)
+     - parameter file: (form) file to upload (optional)
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func uploadFile(petId: Int64, additionalMetadata: String? = nil, file: Data? = nil, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: ApiResponse?, _ error: Error?) -> Void)) {
+        uploadFileWithRequestBuilder(petId: petId, additionalMetadata: additionalMetadata, file: file).execute(apiResponseQueue) { result in
+            switch result {
+            case let .success(response):
+                completion(response.body, nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     uploads an image
+     - POST /pet/{petId}/uploadImage
+     - OAuth:
+       - type: oauth2
+       - name: petstore_auth
+     - parameter petId: (path) ID of pet to update
+     - parameter additionalMetadata: (form) Additional data to pass to server (optional)
+     - parameter file: (form) file to upload (optional)
+     - returns: RequestBuilder<ApiResponse>
+     */
+    open class func uploadFileWithRequestBuilder(petId: Int64, additionalMetadata: String? = nil, file: Data? = nil) -> RequestBuilder<ApiResponse> {
+        var localVariablePath = "/pet/{petId}/uploadImage"
+        let petIdPreEscape = "\(APIHelper.mapValueToPathItem(petId))"
+        let petIdPostEscape = petIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
+        localVariablePath = localVariablePath.replacingOccurrences(of: "{petId}", with: petIdPostEscape, options: .literal, range: nil)
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableFormParams: [String: Any?] = [
+            "additionalMetadata": additionalMetadata?.encodeToJSON(),
+            "file": file?.encodeToJSON(),
+        ]
+
+        let localVariableNonNullParameters = APIHelper.rejectNil(localVariableFormParams)
+        let localVariableParameters = APIHelper.convertBoolToString(localVariableNonNullParameters)
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            "Content-Type": "multipart/form-data",
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<ApiResponse>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
+
+        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+}
+
+
+//
+// UserAPI.swift
+//
+// Generated by openapi-generator
+// https://openapi-generator.tech
+//
+
+import Foundation
+import Combine
+// import Runtime
+
+
+open class UserAPI {
+    private let transport: Transport
+
+    public init(_ transport: Transport) {
+        self.transport = transport
+    }
+    /**
+        Create user
+    - POST /user
+        - This can only be done by the logged in user.
+        - API Key:
+        - type: apiKey api_key
+        - name: api_key
+        - parameter user: (body) Created user object
+    - returns: Future<Void, Never>
+        */
+    open func createUser(user: User) -> Future<Void, Never> {
+    }
+    /**
+        Creates list of users with given input array
+    - POST /user/createWithArray
+        - API Key:
+        - type: apiKey api_key
+        - name: api_key
+        - parameter user: (body) List of user object
+    - returns: Future<Void, Never>
+        */
+    open func createUsersWithArrayInput(user: [User]) -> Future<Void, Never> {
+    }
+    /**
+        Creates list of users with given input array
+    - POST /user/createWithList
+        - API Key:
+        - type: apiKey api_key
+        - name: api_key
+        - parameter user: (body) List of user object
+    - returns: Future<Void, Never>
+        */
+    open func createUsersWithListInput(user: [User]) -> Future<Void, Never> {
+    }
+    /**
+        Delete user
+    - DELETE /user/{username}
+        - This can only be done by the logged in user.
+        - API Key:
+        - type: apiKey api_key
+        - name: api_key
+        - parameter username: (path) The name that needs to be deleted
+    - returns: Future<Void, Never>
+        */
+    open func deleteUser(username: String) -> Future<Void, Never> {
+    }
+    /**
+        Get user by user name
+    - GET /user/{username}
+        - parameter username: (path) The name that needs to be fetched. Use user1 for testing.
+    - returns: Future<User, Never>
+        */
+    open func getUserByName(username: String) -> Future<User, Never> {
+    }
+    /**
+        Logs user into the system
+    - GET /user/login
+        - responseHeaders: [Set-Cookie(String), X-Rate-Limit(Int), X-Expires-After(Date)]
+        - parameter username: () The user name for login
+        - parameter password: () The password for login in clear text
+    - returns: Future<String, Never>
+        */
+    open func loginUser(username: String, password: String) -> Future<String, Never> {
+    }
+    /**
+        Logs out current logged in user session
+    - GET /user/logout
+        - API Key:
+        - type: apiKey api_key
+        - name: api_key
+    - returns: Future<Void, Never>
+        */
+    open func logoutUser() -> Future<Void, Never> {
+    }
+    /**
+        Updated user
+    - PUT /user/{username}
+        - This can only be done by the logged in user.
+        - API Key:
+        - type: apiKey api_key
+        - name: api_key
+        - parameter username: (path) name that need to be deleted
+        - parameter user: (body) Updated user object
+    - returns: Future<Void, Never>
+        */
+    open func updateUser(username: String, user: User) -> Future<Void, Never> {
+    }
+}
+////
+
+/**
+-
+    - API Key:
+    - type: apiKey api_key
+    - name: api_key
+    - OAuth:
+    - type: oauth2
+    - name: petstore_auth
+- returns: RequestBuilder<Void>
+    */
+open class func WithRequestBuilder() -> RequestBuilder<Void> {
+let localVariablePath = ""
+    let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+            let localVariableParameters: [String: Any]? = nil
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+    let localVariableNillableHeaders: [String: Any?] = [
+        :
+    ]
+
+    let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+    let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+    return localVariableRequestBuilder.init(method: "", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+import Foundation
+#if canImport(AnyCodable)
+import AnyCodable
+#endif
+
+open class UserAPI {
+
+    /**
+     Create user
+     
+     - parameter user: (body) Created user object
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func createUser(user: User, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
+        createUserWithRequestBuilder(user: user).execute(apiResponseQueue) { result in
+            switch result {
+            case .success:
+                completion((), nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Create user
+     - POST /user
+     - This can only be done by the logged in user.
+     - API Key:
+       - type: apiKey api_key
+       - name: api_key
+     - parameter user: (body) Created user object
+     - returns: RequestBuilder<Void>
+     */
+    open class func createUserWithRequestBuilder(user: User) -> RequestBuilder<Void> {
+        let localVariablePath = "/user"
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: user)
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Creates list of users with given input array
+     
+     - parameter user: (body) List of user object
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func createUsersWithArrayInput(user: [User], apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
+        createUsersWithArrayInputWithRequestBuilder(user: user).execute(apiResponseQueue) { result in
+            switch result {
+            case .success:
+                completion((), nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Creates list of users with given input array
+     - POST /user/createWithArray
+     - API Key:
+       - type: apiKey api_key
+       - name: api_key
+     - parameter user: (body) List of user object
+     - returns: RequestBuilder<Void>
+     */
+    open class func createUsersWithArrayInputWithRequestBuilder(user: [User]) -> RequestBuilder<Void> {
+        let localVariablePath = "/user/createWithArray"
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: user)
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Creates list of users with given input array
+     
+     - parameter user: (body) List of user object
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func createUsersWithListInput(user: [User], apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
+        createUsersWithListInputWithRequestBuilder(user: user).execute(apiResponseQueue) { result in
+            switch result {
+            case .success:
+                completion((), nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Creates list of users with given input array
+     - POST /user/createWithList
+     - API Key:
+       - type: apiKey api_key
+       - name: api_key
+     - parameter user: (body) List of user object
+     - returns: RequestBuilder<Void>
+     */
+    open class func createUsersWithListInputWithRequestBuilder(user: [User]) -> RequestBuilder<Void> {
+        let localVariablePath = "/user/createWithList"
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: user)
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Delete user
+     
+     - parameter username: (path) The name that needs to be deleted
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func deleteUser(username: String, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
+        deleteUserWithRequestBuilder(username: username).execute(apiResponseQueue) { result in
+            switch result {
+            case .success:
+                completion((), nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Delete user
+     - DELETE /user/{username}
+     - This can only be done by the logged in user.
+     - API Key:
+       - type: apiKey api_key
+       - name: api_key
+     - parameter username: (path) The name that needs to be deleted
+     - returns: RequestBuilder<Void>
+     */
+    open class func deleteUserWithRequestBuilder(username: String) -> RequestBuilder<Void> {
+        var localVariablePath = "/user/{username}"
+        let usernamePreEscape = "\(APIHelper.mapValueToPathItem(username))"
+        let usernamePostEscape = usernamePreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
+        localVariablePath = localVariablePath.replacingOccurrences(of: "{username}", with: usernamePostEscape, options: .literal, range: nil)
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters: [String: Any]? = nil
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+        return localVariableRequestBuilder.init(method: "DELETE", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Get user by user name
+     
+     - parameter username: (path) The name that needs to be fetched. Use user1 for testing.
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func getUserByName(username: String, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: User?, _ error: Error?) -> Void)) {
+        getUserByNameWithRequestBuilder(username: username).execute(apiResponseQueue) { result in
+            switch result {
+            case let .success(response):
+                completion(response.body, nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Get user by user name
+     - GET /user/{username}
+     - parameter username: (path) The name that needs to be fetched. Use user1 for testing.
+     - returns: RequestBuilder<User>
+     */
+    open class func getUserByNameWithRequestBuilder(username: String) -> RequestBuilder<User> {
+        var localVariablePath = "/user/{username}"
+        let usernamePreEscape = "\(APIHelper.mapValueToPathItem(username))"
+        let usernamePostEscape = usernamePreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
+        localVariablePath = localVariablePath.replacingOccurrences(of: "{username}", with: usernamePostEscape, options: .literal, range: nil)
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters: [String: Any]? = nil
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<User>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
+
+        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Logs user into the system
+     
+     - parameter username: (query) The user name for login
+     - parameter password: (query) The password for login in clear text
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func loginUser(username: String, password: String, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: String?, _ error: Error?) -> Void)) {
+        loginUserWithRequestBuilder(username: username, password: password).execute(apiResponseQueue) { result in
+            switch result {
+            case let .success(response):
+                completion(response.body, nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Logs user into the system
+     - GET /user/login
+     - responseHeaders: [Set-Cookie(String), X-Rate-Limit(Int), X-Expires-After(Date)]
+     - parameter username: (query) The user name for login
+     - parameter password: (query) The password for login in clear text
+     - returns: RequestBuilder<String>
+     */
+    open class func loginUserWithRequestBuilder(username: String, password: String) -> RequestBuilder<String> {
+        let localVariablePath = "/user/login"
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters: [String: Any]? = nil
+
+        var localVariableUrlComponents = URLComponents(string: localVariableURLString)
+        localVariableUrlComponents?.queryItems = APIHelper.mapValuesToQueryItems([
+            "username": username.encodeToJSON(),
+            "password": password.encodeToJSON(),
+        ])
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<String>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
+
+        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Logs out current logged in user session
+     
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func logoutUser(apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
+        logoutUserWithRequestBuilder().execute(apiResponseQueue) { result in
+            switch result {
+            case .success:
+                completion((), nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Logs out current logged in user session
+     - GET /user/logout
+     - API Key:
+       - type: apiKey api_key
+       - name: api_key
+     - returns: RequestBuilder<Void>
+     */
+    open class func logoutUserWithRequestBuilder() -> RequestBuilder<Void> {
+        let localVariablePath = "/user/logout"
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters: [String: Any]? = nil
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Updated user
+     
+     - parameter username: (path) name that need to be deleted
+     - parameter user: (body) Updated user object
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func updateUser(username: String, user: User, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
+        updateUserWithRequestBuilder(username: username, user: user).execute(apiResponseQueue) { result in
+            switch result {
+            case .success:
+                completion((), nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Updated user
+     - PUT /user/{username}
+     - This can only be done by the logged in user.
+     - API Key:
+       - type: apiKey api_key
+       - name: api_key
+     - parameter username: (path) name that need to be deleted
+     - parameter user: (body) Updated user object
+     - returns: RequestBuilder<Void>
+     */
+    open class func updateUserWithRequestBuilder(username: String, user: User) -> RequestBuilder<Void> {
+        var localVariablePath = "/user/{username}"
+        let usernamePreEscape = "\(APIHelper.mapValueToPathItem(username))"
+        let usernamePostEscape = usernamePreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
+        localVariablePath = localVariablePath.replacingOccurrences(of: "{username}", with: usernamePostEscape, options: .literal, range: nil)
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: user)
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+        return localVariableRequestBuilder.init(method: "PUT", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+}
+
+
+//
+// StoreAPI.swift
+//
+// Generated by openapi-generator
+// https://openapi-generator.tech
+//
+
+import Foundation
+import Combine
+// import Runtime
+
+
+open class StoreAPI {
+    private let transport: Transport
+
+    public init(_ transport: Transport) {
+        self.transport = transport
+    }
+    /**
+        Delete purchase order by ID
+    - DELETE /store/order/{orderId}
+        - For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors
+        - parameter orderId: (path) ID of the order that needs to be deleted
+    - returns: Future<Void, Never>
+        */
+    open func deleteOrder(orderId: String) -> Future<Void, Never> {
+    }
+    /**
+        Returns pet inventories by status
+    - GET /store/inventory
+        - Returns a map of status codes to quantities
+        - API Key:
+        - type: apiKey api_key
+        - name: api_key
+    - returns: Future<[String: Int], Never>
+        */
+    open func getInventory() -> Future<[String: Int], Never> {
+    }
+    /**
+        Find purchase order by ID
+    - GET /store/order/{orderId}
+        - For valid response try integer IDs with value <= 5 or > 10. Other values will generated exceptions
+        - parameter orderId: (path) ID of pet that needs to be fetched
+    - returns: Future<Order, Never>
+        */
+    open func getOrderById(orderId: Int64) -> Future<Order, Never> {
+    }
+    /**
+        Place an order for a pet
+    - POST /store/order
+        - parameter order: (body) order placed for purchasing the pet
+    - returns: Future<Order, Never>
+        */
+    open func placeOrder(order: Order) -> Future<Order, Never> {
+    }
+}
+////
+
+/**
+-
+    - API Key:
+    - type: apiKey api_key
+    - name: api_key
+    - OAuth:
+    - type: oauth2
+    - name: petstore_auth
+- returns: RequestBuilder<Void>
+    */
+open class func WithRequestBuilder() -> RequestBuilder<Void> {
+let localVariablePath = ""
+    let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+            let localVariableParameters: [String: Any]? = nil
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+    let localVariableNillableHeaders: [String: Any?] = [
+        :
+    ]
+
+    let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+    let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+    return localVariableRequestBuilder.init(method: "", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+import Foundation
+#if canImport(AnyCodable)
+import AnyCodable
+#endif
+
+open class StoreAPI {
+
+    /**
+     Delete purchase order by ID
+     
+     - parameter orderId: (path) ID of the order that needs to be deleted
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func deleteOrder(orderId: String, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
+        deleteOrderWithRequestBuilder(orderId: orderId).execute(apiResponseQueue) { result in
+            switch result {
+            case .success:
+                completion((), nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Delete purchase order by ID
+     - DELETE /store/order/{orderId}
+     - For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors
+     - parameter orderId: (path) ID of the order that needs to be deleted
+     - returns: RequestBuilder<Void>
+     */
+    open class func deleteOrderWithRequestBuilder(orderId: String) -> RequestBuilder<Void> {
+        var localVariablePath = "/store/order/{orderId}"
+        let orderIdPreEscape = "\(APIHelper.mapValueToPathItem(orderId))"
+        let orderIdPostEscape = orderIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
+        localVariablePath = localVariablePath.replacingOccurrences(of: "{orderId}", with: orderIdPostEscape, options: .literal, range: nil)
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters: [String: Any]? = nil
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
+
+        return localVariableRequestBuilder.init(method: "DELETE", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Returns pet inventories by status
+     
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func getInventory(apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: [String: Int]?, _ error: Error?) -> Void)) {
+        getInventoryWithRequestBuilder().execute(apiResponseQueue) { result in
+            switch result {
+            case let .success(response):
+                completion(response.body, nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Returns pet inventories by status
+     - GET /store/inventory
+     - Returns a map of status codes to quantities
+     - API Key:
+       - type: apiKey api_key
+       - name: api_key
+     - returns: RequestBuilder<[String: Int]>
+     */
+    open class func getInventoryWithRequestBuilder() -> RequestBuilder<[String: Int]> {
+        let localVariablePath = "/store/inventory"
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters: [String: Any]? = nil
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<[String: Int]>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
+
+        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Find purchase order by ID
+     
+     - parameter orderId: (path) ID of pet that needs to be fetched
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func getOrderById(orderId: Int64, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Order?, _ error: Error?) -> Void)) {
+        getOrderByIdWithRequestBuilder(orderId: orderId).execute(apiResponseQueue) { result in
+            switch result {
+            case let .success(response):
+                completion(response.body, nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Find purchase order by ID
+     - GET /store/order/{orderId}
+     - For valid response try integer IDs with value <= 5 or > 10. Other values will generated exceptions
+     - parameter orderId: (path) ID of pet that needs to be fetched
+     - returns: RequestBuilder<Order>
+     */
+    open class func getOrderByIdWithRequestBuilder(orderId: Int64) -> RequestBuilder<Order> {
+        var localVariablePath = "/store/order/{orderId}"
+        let orderIdPreEscape = "\(APIHelper.mapValueToPathItem(orderId))"
+        let orderIdPostEscape = orderIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
+        localVariablePath = localVariablePath.replacingOccurrences(of: "{orderId}", with: orderIdPostEscape, options: .literal, range: nil)
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters: [String: Any]? = nil
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Order>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
+
+        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+
+    /**
+     Place an order for a pet
+     
+     - parameter order: (body) order placed for purchasing the pet
+     - parameter apiResponseQueue: The queue on which api response is dispatched.
+     - parameter completion: completion handler to receive the data and the error objects
+     */
+    open class func placeOrder(order: Order, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Order?, _ error: Error?) -> Void)) {
+        placeOrderWithRequestBuilder(order: order).execute(apiResponseQueue) { result in
+            switch result {
+            case let .success(response):
+                completion(response.body, nil)
+            case let .failure(error):
+                completion(nil, error)
+            }
+        }
+    }
+
+    /**
+     Place an order for a pet
+     - POST /store/order
+     - parameter order: (body) order placed for purchasing the pet
+     - returns: RequestBuilder<Order>
+     */
+    open class func placeOrderWithRequestBuilder(order: Order) -> RequestBuilder<Order> {
+        let localVariablePath = "/store/order"
+        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
+        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: order)
+
+        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+
+        let localVariableNillableHeaders: [String: Any?] = [
+            :
+        ]
+
+        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
+
+        let localVariableRequestBuilder: RequestBuilder<Order>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
+
+        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    }
+}
