+++ modules/openapi-generator/src/main/resources/swift-alt/_converted_param.mustache
@@ -1,0 +1,1 @@
+{{#isArray}}{{paramName}}{{^required}}?{{/required}}.joined(separator: ", "){{/isArray}}{{^isArray}}{{paramName}}{{/isArray}}
