+++ samples/client/petstore/swift/alt/OpenAPIClient/Classes/OpenAPIs/APIs/PetAPI.swift
@@ -3,7 +3,7 @@
 //
 // Generated by openapi-generator
 // https://openapi-generator.tech
-//
+// Run in root folder to build example mvn clean package -DskipTests; ./bin/generate-samples.sh bin/configs/swift-alt-petstore-new.yaml
 
 import Foundation
 import Combine
@@ -11,556 +11,381 @@ import Combine
 
 
 open class PetAPI {
+    private let encoder: JSONEncoder = {
+        let encoder = JSONEncoder()
+        encoder.dateEncodingStrategy = .formatted(OpenISO8601DateFormatter())
+        encoder.outputFormatting = .prettyPrinted
+        return encoder
+    }()
+    private let decoder: JSONDecoder = {
+        let decoder = JSONDecoder()
+        decoder.dateDecodingStrategy = .formatted(OpenISO8601DateFormatter())
+        return decoder
+    }()
     private let transport: Transport
 
     public init(_ transport: Transport) {
         self.transport = transport
     }
+
     /**
         Add a new pet to the store
-    - POST /pet
+        - POST /pet
         - OAuth:
         - type: oauth2
         - name: petstore_auth
         - parameter pet: (body) Pet object that needs to be added to the store 
-    - returns: Future<Pet, Never> 
+        - returns: AnyPublisher<Result<Pet, Error>, Never> 
         */
-    open func addPet(pet: Pet) -> Future<Pet, Never> {
+    open func addPet(pet: Pet) -> AnyPublisher<Result<Pet, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/pet"
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "POST"
+        
+        // Setting body parameters 
+        request.httpBody = try? encoder.encode(pet)
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Pet, Error> in
+                if response.response?.statusCode == 200 {
+                    if let data = response.data {
+                        return Result { try self.decoder.decode(Pet.self, from: data) }
+                    } else {
+                        return .failure(TransportError.unknown)
+                    }
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Deletes a pet
-    - DELETE /pet/{petId}
+        - DELETE /pet/{petId}
         - OAuth:
         - type: oauth2
         - name: petstore_auth
         - parameter petId: (path) Pet id to delete 
         - parameter apiKey: (header)  (optional)
-    - returns: Future<Void, Never> 
+        - returns: AnyPublisher<Result<Void, Error>, Never> 
         */
-    open func deletePet(petId: Int64, apiKey: String? = nil) -> Future<Void, Never> {
+    open func deletePet(petId: Int64, apiKey: String? = nil) -> AnyPublisher<Result<Void, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/pet/{petId}"
+            .replacingOccurrences(of: "{petId}", with: "\(petId)")
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "DELETE"
+        // Setting headers
+        request.allHTTPHeaderFields = [
+            "api_key": apiKey
+        ].compactMapValues { $0 }
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Void, Error> in
+                if response.response?.statusCode == 200 {
+                    return .success(())
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Finds Pets by status
-    - GET /pet/findByStatus
+        - GET /pet/findByStatus
         - Multiple status values can be provided with comma separated strings
         - OAuth:
         - type: oauth2
         - name: petstore_auth
-        - parameter status: () Status values that need to be considered for filter 
-    - returns: Future<[Pet], Never> 
+        - parameter status: (query) Status values that need to be considered for filter 
+        - returns: AnyPublisher<Result<[Pet], Error>, Never> 
         */
-    open func findPetsByStatus(status: [String]) -> Future<[Pet], Never> {
+    open func findPetsByStatus(status: [String]) -> AnyPublisher<Result<[Pet], Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/pet/findByStatus"
+        components.queryItems = [
+            URLQueryItem(name: "status", value: status.joined(separator: ", "))
+        ]
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "GET"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<[Pet], Error> in
+                if response.response?.statusCode == 200 {
+                    if let data = response.data {
+                        return Result { try self.decoder.decode([Pet].self, from: data) }
+                    } else {
+                        return .failure(TransportError.unknown)
+                    }
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Finds Pets by tags
-    - GET /pet/findByTags
+        - GET /pet/findByTags
         - Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.
         - OAuth:
         - type: oauth2
         - name: petstore_auth
-        - parameter tags: () Tags to filter by 
-    - returns: Future<[Pet], Never> 
+        - parameter tags: (query) Tags to filter by 
+        - returns: AnyPublisher<Result<[Pet], Error>, Never> 
         */
-    open func findPetsByTags(tags: [String]) -> Future<[Pet], Never> {
+    open func findPetsByTags(tags: [String]) -> AnyPublisher<Result<[Pet], Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/pet/findByTags"
+        components.queryItems = [
+            URLQueryItem(name: "tags", value: tags.joined(separator: ", "))
+        ]
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "GET"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<[Pet], Error> in
+                if response.response?.statusCode == 200 {
+                    if let data = response.data {
+                        return Result { try self.decoder.decode([Pet].self, from: data) }
+                    } else {
+                        return .failure(TransportError.unknown)
+                    }
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Find pet by ID
-    - GET /pet/{petId}
+        - GET /pet/{petId}
         - Returns a single pet
         - API Key:
         - type: apiKey api_key 
         - name: api_key
         - parameter petId: (path) ID of pet to return 
-    - returns: Future<Pet, Never> 
+        - returns: AnyPublisher<Result<Pet, Error>, Never> 
         */
-    open func getPetById(petId: Int64) -> Future<Pet, Never> {
+    open func getPetById(petId: Int64) -> AnyPublisher<Result<Pet, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/pet/{petId}"
+            .replacingOccurrences(of: "{petId}", with: "\(petId)")
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "GET"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Pet, Error> in
+                if response.response?.statusCode == 200 {
+                    if let data = response.data {
+                        return Result { try self.decoder.decode(Pet.self, from: data) }
+                    } else {
+                        return .failure(TransportError.unknown)
+                    }
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Update an existing pet
-    - PUT /pet
+        - PUT /pet
         - OAuth:
         - type: oauth2
         - name: petstore_auth
         - parameter pet: (body) Pet object that needs to be added to the store 
-    - returns: Future<Pet, Never> 
+        - returns: AnyPublisher<Result<Pet, Error>, Never> 
         */
-    open func updatePet(pet: Pet) -> Future<Pet, Never> {
+    open func updatePet(pet: Pet) -> AnyPublisher<Result<Pet, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/pet"
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "PUT"
+        
+        // Setting body parameters 
+        request.httpBody = try? encoder.encode(pet)
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Pet, Error> in
+                if response.response?.statusCode == 200 {
+                    if let data = response.data {
+                        return Result { try self.decoder.decode(Pet.self, from: data) }
+                    } else {
+                        return .failure(TransportError.unknown)
+                    }
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Updates a pet in the store with form data
-    - POST /pet/{petId}
+        - POST /pet/{petId}
         - OAuth:
         - type: oauth2
         - name: petstore_auth
         - parameter petId: (path) ID of pet that needs to be updated 
         - parameter name: (form) Updated name of the pet (optional)
         - parameter status: (form) Updated status of the pet (optional)
-    - returns: Future<Void, Never> 
+        - returns: AnyPublisher<Result<Void, Error>, Never> 
         */
-    open func updatePetWithForm(petId: Int64, name: String? = nil, status: String? = nil) -> Future<Void, Never> {
+    open func updatePetWithForm(petId: Int64, name: String? = nil, status: String? = nil) -> AnyPublisher<Result<Void, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/pet/{petId}"
+            .replacingOccurrences(of: "{petId}", with: "\(petId)")
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "POST"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Void, Error> in
+                if response.response?.statusCode == 200 {
+                    return .success(())
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         uploads an image
-    - POST /pet/{petId}/uploadImage
+        - POST /pet/{petId}/uploadImage
         - OAuth:
         - type: oauth2
         - name: petstore_auth
         - parameter petId: (path) ID of pet to update 
         - parameter additionalMetadata: (form) Additional data to pass to server (optional)
         - parameter file: (form) file to upload (optional)
-    - returns: Future<ApiResponse, Never> 
+        - returns: AnyPublisher<Result<ApiResponse, Error>, Never> 
         */
-    open func uploadFile(petId: Int64, additionalMetadata: String? = nil, file: Data? = nil) -> Future<ApiResponse, Never> {
+    open func uploadFile(petId: Int64, additionalMetadata: String? = nil, file: Data? = nil) -> AnyPublisher<Result<ApiResponse, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/pet/{petId}/uploadImage"
+            .replacingOccurrences(of: "{petId}", with: "\(petId)")
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "POST"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<ApiResponse, Error> in
+                if response.response?.statusCode == 200 {
+                    if let data = response.data {
+                        return Result { try self.decoder.decode(ApiResponse.self, from: data) }
+                    } else {
+                        return .failure(TransportError.unknown)
+                    }
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
 }
-////
 
-/**
--  
-    - API Key:
-    - type: apiKey api_key 
-    - name: api_key
-    - OAuth:
-    - type: oauth2
-    - name: petstore_auth
-- returns: RequestBuilder<Void> 
-    */
-open class func WithRequestBuilder() -> RequestBuilder<Void> {
-let localVariablePath = ""
-    let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-            let localVariableParameters: [String: Any]? = nil
 
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+private class OpenISO8601DateFormatter: DateFormatter {
+    static let withoutSeconds: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.calendar = Calendar(identifier: .iso8601)
+        formatter.locale = Locale(identifier: "en_US_POSIX")
+        formatter.timeZone = TimeZone(secondsFromGMT: 0)
+        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZZZZZ"
+        return formatter
+    }()
 
-    let localVariableNillableHeaders: [String: Any?] = [
-        :
-    ]
-
-    let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-    let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-    return localVariableRequestBuilder.init(method: "", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    private func setup() {
+        calendar = Calendar(identifier: .iso8601)
+        locale = Locale(identifier: "en_US_POSIX")
+        timeZone = TimeZone(secondsFromGMT: 0)
+        dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZZZZZ"
     }
 
-import Foundation
-#if canImport(AnyCodable)
-import AnyCodable
-#endif
+    override init() {
+        super.init()
+        setup()
+    }
 
-open class PetAPI {
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+        setup()
+    }
 
-    /**
-     Add a new pet to the store
-     
-     - parameter pet: (body) Pet object that needs to be added to the store 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func addPet(pet: Pet, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Pet?, _ error: Error?) -> Void)) {
-        addPetWithRequestBuilder(pet: pet).execute(apiResponseQueue) { result in
-            switch result {
-            case let .success(response):
-                completion(response.body, nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
+    override func date(from string: String) -> Date? {
+        if let result = super.date(from: string) {
+            return result
         }
-    }
-
-    /**
-     Add a new pet to the store
-     - POST /pet
-     - OAuth:
-       - type: oauth2
-       - name: petstore_auth
-     - parameter pet: (body) Pet object that needs to be added to the store 
-     - returns: RequestBuilder<Pet> 
-     */
-    open class func addPetWithRequestBuilder(pet: Pet) -> RequestBuilder<Pet> {
-        let localVariablePath = "/pet"
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: pet)
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Pet>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
-
-        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Deletes a pet
-     
-     - parameter petId: (path) Pet id to delete 
-     - parameter apiKey: (header)  (optional)
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func deletePet(petId: Int64, apiKey: String? = nil, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
-        deletePetWithRequestBuilder(petId: petId, apiKey: apiKey).execute(apiResponseQueue) { result in
-            switch result {
-            case .success:
-                completion((), nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Deletes a pet
-     - DELETE /pet/{petId}
-     - OAuth:
-       - type: oauth2
-       - name: petstore_auth
-     - parameter petId: (path) Pet id to delete 
-     - parameter apiKey: (header)  (optional)
-     - returns: RequestBuilder<Void> 
-     */
-    open class func deletePetWithRequestBuilder(petId: Int64, apiKey: String? = nil) -> RequestBuilder<Void> {
-        var localVariablePath = "/pet/{petId}"
-        let petIdPreEscape = "\(APIHelper.mapValueToPathItem(petId))"
-        let petIdPostEscape = petIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
-        localVariablePath = localVariablePath.replacingOccurrences(of: "{petId}", with: petIdPostEscape, options: .literal, range: nil)
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters: [String: Any]? = nil
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            "api_key": apiKey?.encodeToJSON(),
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-        return localVariableRequestBuilder.init(method: "DELETE", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     * enum for parameter status
-     */
-    public enum Status_findPetsByStatus: String, CaseIterable {
-        case available = "available"
-        case pending = "pending"
-        case sold = "sold"
-    }
-
-    /**
-     Finds Pets by status
-     
-     - parameter status: (query) Status values that need to be considered for filter 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func findPetsByStatus(status: [String], apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: [Pet]?, _ error: Error?) -> Void)) {
-        findPetsByStatusWithRequestBuilder(status: status).execute(apiResponseQueue) { result in
-            switch result {
-            case let .success(response):
-                completion(response.body, nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Finds Pets by status
-     - GET /pet/findByStatus
-     - Multiple status values can be provided with comma separated strings
-     - OAuth:
-       - type: oauth2
-       - name: petstore_auth
-     - parameter status: (query) Status values that need to be considered for filter 
-     - returns: RequestBuilder<[Pet]> 
-     */
-    open class func findPetsByStatusWithRequestBuilder(status: [String]) -> RequestBuilder<[Pet]> {
-        let localVariablePath = "/pet/findByStatus"
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters: [String: Any]? = nil
-
-        var localVariableUrlComponents = URLComponents(string: localVariableURLString)
-        localVariableUrlComponents?.queryItems = APIHelper.mapValuesToQueryItems([
-            "status": status.encodeToJSON(),
-        ])
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<[Pet]>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
-
-        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Finds Pets by tags
-     
-     - parameter tags: (query) Tags to filter by 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    @available(*, deprecated, message: "This operation is deprecated.")
-    open class func findPetsByTags(tags: [String], apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: [Pet]?, _ error: Error?) -> Void)) {
-        findPetsByTagsWithRequestBuilder(tags: tags).execute(apiResponseQueue) { result in
-            switch result {
-            case let .success(response):
-                completion(response.body, nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Finds Pets by tags
-     - GET /pet/findByTags
-     - Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.
-     - OAuth:
-       - type: oauth2
-       - name: petstore_auth
-     - parameter tags: (query) Tags to filter by 
-     - returns: RequestBuilder<[Pet]> 
-     */
-    @available(*, deprecated, message: "This operation is deprecated.")
-    open class func findPetsByTagsWithRequestBuilder(tags: [String]) -> RequestBuilder<[Pet]> {
-        let localVariablePath = "/pet/findByTags"
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters: [String: Any]? = nil
-
-        var localVariableUrlComponents = URLComponents(string: localVariableURLString)
-        localVariableUrlComponents?.queryItems = APIHelper.mapValuesToQueryItems([
-            "tags": tags.encodeToJSON(),
-        ])
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<[Pet]>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
-
-        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Find pet by ID
-     
-     - parameter petId: (path) ID of pet to return 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func getPetById(petId: Int64, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Pet?, _ error: Error?) -> Void)) {
-        getPetByIdWithRequestBuilder(petId: petId).execute(apiResponseQueue) { result in
-            switch result {
-            case let .success(response):
-                completion(response.body, nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Find pet by ID
-     - GET /pet/{petId}
-     - Returns a single pet
-     - API Key:
-       - type: apiKey api_key 
-       - name: api_key
-     - parameter petId: (path) ID of pet to return 
-     - returns: RequestBuilder<Pet> 
-     */
-    open class func getPetByIdWithRequestBuilder(petId: Int64) -> RequestBuilder<Pet> {
-        var localVariablePath = "/pet/{petId}"
-        let petIdPreEscape = "\(APIHelper.mapValueToPathItem(petId))"
-        let petIdPostEscape = petIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
-        localVariablePath = localVariablePath.replacingOccurrences(of: "{petId}", with: petIdPostEscape, options: .literal, range: nil)
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters: [String: Any]? = nil
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Pet>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
-
-        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Update an existing pet
-     
-     - parameter pet: (body) Pet object that needs to be added to the store 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func updatePet(pet: Pet, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Pet?, _ error: Error?) -> Void)) {
-        updatePetWithRequestBuilder(pet: pet).execute(apiResponseQueue) { result in
-            switch result {
-            case let .success(response):
-                completion(response.body, nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Update an existing pet
-     - PUT /pet
-     - OAuth:
-       - type: oauth2
-       - name: petstore_auth
-     - parameter pet: (body) Pet object that needs to be added to the store 
-     - returns: RequestBuilder<Pet> 
-     */
-    open class func updatePetWithRequestBuilder(pet: Pet) -> RequestBuilder<Pet> {
-        let localVariablePath = "/pet"
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: pet)
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Pet>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
-
-        return localVariableRequestBuilder.init(method: "PUT", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Updates a pet in the store with form data
-     
-     - parameter petId: (path) ID of pet that needs to be updated 
-     - parameter name: (form) Updated name of the pet (optional)
-     - parameter status: (form) Updated status of the pet (optional)
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func updatePetWithForm(petId: Int64, name: String? = nil, status: String? = nil, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
-        updatePetWithFormWithRequestBuilder(petId: petId, name: name, status: status).execute(apiResponseQueue) { result in
-            switch result {
-            case .success:
-                completion((), nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Updates a pet in the store with form data
-     - POST /pet/{petId}
-     - OAuth:
-       - type: oauth2
-       - name: petstore_auth
-     - parameter petId: (path) ID of pet that needs to be updated 
-     - parameter name: (form) Updated name of the pet (optional)
-     - parameter status: (form) Updated status of the pet (optional)
-     - returns: RequestBuilder<Void> 
-     */
-    open class func updatePetWithFormWithRequestBuilder(petId: Int64, name: String? = nil, status: String? = nil) -> RequestBuilder<Void> {
-        var localVariablePath = "/pet/{petId}"
-        let petIdPreEscape = "\(APIHelper.mapValueToPathItem(petId))"
-        let petIdPostEscape = petIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
-        localVariablePath = localVariablePath.replacingOccurrences(of: "{petId}", with: petIdPostEscape, options: .literal, range: nil)
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableFormParams: [String: Any?] = [
-            "name": name?.encodeToJSON(),
-            "status": status?.encodeToJSON(),
-        ]
-
-        let localVariableNonNullParameters = APIHelper.rejectNil(localVariableFormParams)
-        let localVariableParameters = APIHelper.convertBoolToString(localVariableNonNullParameters)
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            "Content-Type": "application/x-www-form-urlencoded",
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     uploads an image
-     
-     - parameter petId: (path) ID of pet to update 
-     - parameter additionalMetadata: (form) Additional data to pass to server (optional)
-     - parameter file: (form) file to upload (optional)
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func uploadFile(petId: Int64, additionalMetadata: String? = nil, file: Data? = nil, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: ApiResponse?, _ error: Error?) -> Void)) {
-        uploadFileWithRequestBuilder(petId: petId, additionalMetadata: additionalMetadata, file: file).execute(apiResponseQueue) { result in
-            switch result {
-            case let .success(response):
-                completion(response.body, nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     uploads an image
-     - POST /pet/{petId}/uploadImage
-     - OAuth:
-       - type: oauth2
-       - name: petstore_auth
-     - parameter petId: (path) ID of pet to update 
-     - parameter additionalMetadata: (form) Additional data to pass to server (optional)
-     - parameter file: (form) file to upload (optional)
-     - returns: RequestBuilder<ApiResponse> 
-     */
-    open class func uploadFileWithRequestBuilder(petId: Int64, additionalMetadata: String? = nil, file: Data? = nil) -> RequestBuilder<ApiResponse> {
-        var localVariablePath = "/pet/{petId}/uploadImage"
-        let petIdPreEscape = "\(APIHelper.mapValueToPathItem(petId))"
-        let petIdPostEscape = petIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
-        localVariablePath = localVariablePath.replacingOccurrences(of: "{petId}", with: petIdPostEscape, options: .literal, range: nil)
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableFormParams: [String: Any?] = [
-            "additionalMetadata": additionalMetadata?.encodeToJSON(),
-            "file": file?.encodeToJSON(),
-        ]
-
-        let localVariableNonNullParameters = APIHelper.rejectNil(localVariableFormParams)
-        let localVariableParameters = APIHelper.convertBoolToString(localVariableNonNullParameters)
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            "Content-Type": "multipart/form-data",
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<ApiResponse>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
-
-        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+        return OpenISO8601DateFormatter.withoutSeconds.date(from: string)
     }
 }
