+++ samples/client/petstore/swift/alt/OpenAPIClient/Classes/OpenAPIs/APIs/StoreAPI.swift
@@ -3,7 +3,7 @@
 //
 // Generated by openapi-generator
 // https://openapi-generator.tech
-//
+// Run in root folder to build example mvn clean package -DskipTests; ./bin/generate-samples.sh bin/configs/swift-alt-petstore-new.yaml
 
 import Foundation
 import Combine
@@ -11,261 +11,204 @@ import Combine
 
 
 open class StoreAPI {
+    private let encoder: JSONEncoder = {
+        let encoder = JSONEncoder()
+        encoder.dateEncodingStrategy = .formatted(OpenISO8601DateFormatter())
+        encoder.outputFormatting = .prettyPrinted
+        return encoder
+    }()
+    private let decoder: JSONDecoder = {
+        let decoder = JSONDecoder()
+        decoder.dateDecodingStrategy = .formatted(OpenISO8601DateFormatter())
+        return decoder
+    }()
     private let transport: Transport
 
     public init(_ transport: Transport) {
         self.transport = transport
     }
+
     /**
         Delete purchase order by ID
-    - DELETE /store/order/{orderId}
+        - DELETE /store/order/{orderId}
         - For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors
         - parameter orderId: (path) ID of the order that needs to be deleted 
-    - returns: Future<Void, Never> 
+        - returns: AnyPublisher<Result<Void, Error>, Never> 
         */
-    open func deleteOrder(orderId: String) -> Future<Void, Never> {
+    open func deleteOrder(orderId: String) -> AnyPublisher<Result<Void, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/store/order/{orderId}"
+            .replacingOccurrences(of: "{orderId}", with: "\(orderId)")
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "DELETE"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Void, Error> in
+                if response.response?.statusCode == 200 {
+                    return .success(())
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Returns pet inventories by status
-    - GET /store/inventory
+        - GET /store/inventory
         - Returns a map of status codes to quantities
         - API Key:
         - type: apiKey api_key 
         - name: api_key
-    - returns: Future<[String: Int], Never> 
+        - returns: AnyPublisher<Result<[String: Int], Error>, Never> 
         */
-    open func getInventory() -> Future<[String: Int], Never> {
+    open func getInventory() -> AnyPublisher<Result<[String: Int], Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/store/inventory"
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "GET"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<[String: Int], Error> in
+                if response.response?.statusCode == 200 {
+                    if let data = response.data {
+                        return Result { try self.decoder.decode([String: Int].self, from: data) }
+                    } else {
+                        return .failure(TransportError.unknown)
+                    }
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Find purchase order by ID
-    - GET /store/order/{orderId}
+        - GET /store/order/{orderId}
         - For valid response try integer IDs with value <= 5 or > 10. Other values will generated exceptions
         - parameter orderId: (path) ID of pet that needs to be fetched 
-    - returns: Future<Order, Never> 
+        - returns: AnyPublisher<Result<Order, Error>, Never> 
         */
-    open func getOrderById(orderId: Int64) -> Future<Order, Never> {
+    open func getOrderById(orderId: Int64) -> AnyPublisher<Result<Order, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/store/order/{orderId}"
+            .replacingOccurrences(of: "{orderId}", with: "\(orderId)")
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "GET"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Order, Error> in
+                if response.response?.statusCode == 200 {
+                    if let data = response.data {
+                        return Result { try self.decoder.decode(Order.self, from: data) }
+                    } else {
+                        return .failure(TransportError.unknown)
+                    }
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Place an order for a pet
-    - POST /store/order
+        - POST /store/order
         - parameter order: (body) order placed for purchasing the pet 
-    - returns: Future<Order, Never> 
+        - returns: AnyPublisher<Result<Order, Error>, Never> 
         */
-    open func placeOrder(order: Order) -> Future<Order, Never> {
+    open func placeOrder(order: Order) -> AnyPublisher<Result<Order, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/store/order"
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "POST"
+        
+        // Setting body parameters 
+        request.httpBody = try? encoder.encode(order)
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Order, Error> in
+                if response.response?.statusCode == 200 {
+                    if let data = response.data {
+                        return Result { try self.decoder.decode(Order.self, from: data) }
+                    } else {
+                        return .failure(TransportError.unknown)
+                    }
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
 }
-////
 
-/**
--  
-    - API Key:
-    - type: apiKey api_key 
-    - name: api_key
-    - OAuth:
-    - type: oauth2
-    - name: petstore_auth
-- returns: RequestBuilder<Void> 
-    */
-open class func WithRequestBuilder() -> RequestBuilder<Void> {
-let localVariablePath = ""
-    let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-            let localVariableParameters: [String: Any]? = nil
 
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+private class OpenISO8601DateFormatter: DateFormatter {
+    static let withoutSeconds: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.calendar = Calendar(identifier: .iso8601)
+        formatter.locale = Locale(identifier: "en_US_POSIX")
+        formatter.timeZone = TimeZone(secondsFromGMT: 0)
+        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZZZZZ"
+        return formatter
+    }()
 
-    let localVariableNillableHeaders: [String: Any?] = [
-        :
-    ]
-
-    let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-    let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-    return localVariableRequestBuilder.init(method: "", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    private func setup() {
+        calendar = Calendar(identifier: .iso8601)
+        locale = Locale(identifier: "en_US_POSIX")
+        timeZone = TimeZone(secondsFromGMT: 0)
+        dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZZZZZ"
     }
 
-import Foundation
-#if canImport(AnyCodable)
-import AnyCodable
-#endif
+    override init() {
+        super.init()
+        setup()
+    }
 
-open class StoreAPI {
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+        setup()
+    }
 
-    /**
-     Delete purchase order by ID
-     
-     - parameter orderId: (path) ID of the order that needs to be deleted 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func deleteOrder(orderId: String, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
-        deleteOrderWithRequestBuilder(orderId: orderId).execute(apiResponseQueue) { result in
-            switch result {
-            case .success:
-                completion((), nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
+    override func date(from string: String) -> Date? {
+        if let result = super.date(from: string) {
+            return result
         }
-    }
-
-    /**
-     Delete purchase order by ID
-     - DELETE /store/order/{orderId}
-     - For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors
-     - parameter orderId: (path) ID of the order that needs to be deleted 
-     - returns: RequestBuilder<Void> 
-     */
-    open class func deleteOrderWithRequestBuilder(orderId: String) -> RequestBuilder<Void> {
-        var localVariablePath = "/store/order/{orderId}"
-        let orderIdPreEscape = "\(APIHelper.mapValueToPathItem(orderId))"
-        let orderIdPostEscape = orderIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
-        localVariablePath = localVariablePath.replacingOccurrences(of: "{orderId}", with: orderIdPostEscape, options: .literal, range: nil)
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters: [String: Any]? = nil
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-        return localVariableRequestBuilder.init(method: "DELETE", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Returns pet inventories by status
-     
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func getInventory(apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: [String: Int]?, _ error: Error?) -> Void)) {
-        getInventoryWithRequestBuilder().execute(apiResponseQueue) { result in
-            switch result {
-            case let .success(response):
-                completion(response.body, nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Returns pet inventories by status
-     - GET /store/inventory
-     - Returns a map of status codes to quantities
-     - API Key:
-       - type: apiKey api_key 
-       - name: api_key
-     - returns: RequestBuilder<[String: Int]> 
-     */
-    open class func getInventoryWithRequestBuilder() -> RequestBuilder<[String: Int]> {
-        let localVariablePath = "/store/inventory"
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters: [String: Any]? = nil
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<[String: Int]>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
-
-        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Find purchase order by ID
-     
-     - parameter orderId: (path) ID of pet that needs to be fetched 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func getOrderById(orderId: Int64, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Order?, _ error: Error?) -> Void)) {
-        getOrderByIdWithRequestBuilder(orderId: orderId).execute(apiResponseQueue) { result in
-            switch result {
-            case let .success(response):
-                completion(response.body, nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Find purchase order by ID
-     - GET /store/order/{orderId}
-     - For valid response try integer IDs with value <= 5 or > 10. Other values will generated exceptions
-     - parameter orderId: (path) ID of pet that needs to be fetched 
-     - returns: RequestBuilder<Order> 
-     */
-    open class func getOrderByIdWithRequestBuilder(orderId: Int64) -> RequestBuilder<Order> {
-        var localVariablePath = "/store/order/{orderId}"
-        let orderIdPreEscape = "\(APIHelper.mapValueToPathItem(orderId))"
-        let orderIdPostEscape = orderIdPreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
-        localVariablePath = localVariablePath.replacingOccurrences(of: "{orderId}", with: orderIdPostEscape, options: .literal, range: nil)
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters: [String: Any]? = nil
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Order>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
-
-        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Place an order for a pet
-     
-     - parameter order: (body) order placed for purchasing the pet 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func placeOrder(order: Order, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Order?, _ error: Error?) -> Void)) {
-        placeOrderWithRequestBuilder(order: order).execute(apiResponseQueue) { result in
-            switch result {
-            case let .success(response):
-                completion(response.body, nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Place an order for a pet
-     - POST /store/order
-     - parameter order: (body) order placed for purchasing the pet 
-     - returns: RequestBuilder<Order> 
-     */
-    open class func placeOrderWithRequestBuilder(order: Order) -> RequestBuilder<Order> {
-        let localVariablePath = "/store/order"
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: order)
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Order>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
-
-        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+        return OpenISO8601DateFormatter.withoutSeconds.date(from: string)
     }
 }
