+++ samples/client/petstore/swift/alt/OpenAPIClient/Classes/OpenAPIs/APIs/UserAPI.swift
@@ -3,7 +3,7 @@
 //
 // Generated by openapi-generator
 // https://openapi-generator.tech
-//
+// Run in root folder to build example mvn clean package -DskipTests; ./bin/generate-samples.sh bin/configs/swift-alt-petstore-new.yaml
 
 import Foundation
 import Combine
@@ -11,506 +11,353 @@ import Combine
 
 
 open class UserAPI {
+    private let encoder: JSONEncoder = {
+        let encoder = JSONEncoder()
+        encoder.dateEncodingStrategy = .formatted(OpenISO8601DateFormatter())
+        encoder.outputFormatting = .prettyPrinted
+        return encoder
+    }()
+    private let decoder: JSONDecoder = {
+        let decoder = JSONDecoder()
+        decoder.dateDecodingStrategy = .formatted(OpenISO8601DateFormatter())
+        return decoder
+    }()
     private let transport: Transport
 
     public init(_ transport: Transport) {
         self.transport = transport
     }
+
     /**
         Create user
-    - POST /user
+        - POST /user
         - This can only be done by the logged in user.
         - API Key:
         - type: apiKey api_key 
         - name: api_key
         - parameter user: (body) Created user object 
-    - returns: Future<Void, Never> 
+        - returns: AnyPublisher<Result<Void, Error>, Never> 
         */
-    open func createUser(user: User) -> Future<Void, Never> {
+    open func createUser(user: User) -> AnyPublisher<Result<Void, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/user"
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "POST"
+        
+        // Setting body parameters 
+        request.httpBody = try? encoder.encode(user)
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Void, Error> in
+                if response.response?.statusCode == 200 {
+                    return .success(())
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Creates list of users with given input array
-    - POST /user/createWithArray
+        - POST /user/createWithArray
         - API Key:
         - type: apiKey api_key 
         - name: api_key
         - parameter user: (body) List of user object 
-    - returns: Future<Void, Never> 
+        - returns: AnyPublisher<Result<Void, Error>, Never> 
         */
-    open func createUsersWithArrayInput(user: [User]) -> Future<Void, Never> {
+    open func createUsersWithArrayInput(user: [User]) -> AnyPublisher<Result<Void, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/user/createWithArray"
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "POST"
+        
+        // Setting body parameters 
+        request.httpBody = try? encoder.encode(user)
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Void, Error> in
+                if response.response?.statusCode == 200 {
+                    return .success(())
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Creates list of users with given input array
-    - POST /user/createWithList
+        - POST /user/createWithList
         - API Key:
         - type: apiKey api_key 
         - name: api_key
         - parameter user: (body) List of user object 
-    - returns: Future<Void, Never> 
+        - returns: AnyPublisher<Result<Void, Error>, Never> 
         */
-    open func createUsersWithListInput(user: [User]) -> Future<Void, Never> {
+    open func createUsersWithListInput(user: [User]) -> AnyPublisher<Result<Void, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/user/createWithList"
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "POST"
+        
+        // Setting body parameters 
+        request.httpBody = try? encoder.encode(user)
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Void, Error> in
+                if response.response?.statusCode == 200 {
+                    return .success(())
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Delete user
-    - DELETE /user/{username}
+        - DELETE /user/{username}
         - This can only be done by the logged in user.
         - API Key:
         - type: apiKey api_key 
         - name: api_key
         - parameter username: (path) The name that needs to be deleted 
-    - returns: Future<Void, Never> 
+        - returns: AnyPublisher<Result<Void, Error>, Never> 
         */
-    open func deleteUser(username: String) -> Future<Void, Never> {
+    open func deleteUser(username: String) -> AnyPublisher<Result<Void, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/user/{username}"
+            .replacingOccurrences(of: "{username}", with: "\(username)")
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "DELETE"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Void, Error> in
+                if response.response?.statusCode == 200 {
+                    return .success(())
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Get user by user name
-    - GET /user/{username}
+        - GET /user/{username}
         - parameter username: (path) The name that needs to be fetched. Use user1 for testing. 
-    - returns: Future<User, Never> 
+        - returns: AnyPublisher<Result<User, Error>, Never> 
         */
-    open func getUserByName(username: String) -> Future<User, Never> {
+    open func getUserByName(username: String) -> AnyPublisher<Result<User, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/user/{username}"
+            .replacingOccurrences(of: "{username}", with: "\(username)")
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "GET"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<User, Error> in
+                if response.response?.statusCode == 200 {
+                    if let data = response.data {
+                        return Result { try self.decoder.decode(User.self, from: data) }
+                    } else {
+                        return .failure(TransportError.unknown)
+                    }
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Logs user into the system
-    - GET /user/login
+        - GET /user/login
         - responseHeaders: [Set-Cookie(String), X-Rate-Limit(Int), X-Expires-After(Date)]
-        - parameter username: () The user name for login 
-        - parameter password: () The password for login in clear text 
-    - returns: Future<String, Never> 
+        - parameter username: (query) The user name for login 
+        - parameter password: (query) The password for login in clear text 
+        - returns: AnyPublisher<Result<String, Error>, Never> 
         */
-    open func loginUser(username: String, password: String) -> Future<String, Never> {
+    open func loginUser(username: String, password: String) -> AnyPublisher<Result<String, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/user/login"
+        components.queryItems = [
+            URLQueryItem(name: "username", value: username), 
+            URLQueryItem(name: "password", value: password)
+        ]
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "GET"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<String, Error> in
+                if response.response?.statusCode == 200 {
+                    if let data = response.data {
+                        return Result { try self.decoder.decode(String.self, from: data) }
+                    } else {
+                        return .failure(TransportError.unknown)
+                    }
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Logs out current logged in user session
-    - GET /user/logout
+        - GET /user/logout
         - API Key:
         - type: apiKey api_key 
         - name: api_key
-    - returns: Future<Void, Never> 
+        - returns: AnyPublisher<Result<Void, Error>, Never> 
         */
-    open func logoutUser() -> Future<Void, Never> {
+    open func logoutUser() -> AnyPublisher<Result<Void, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/user/logout"
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "GET"
+        
+        
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Void, Error> in
+                if response.response?.statusCode == 200 {
+                    return .success(())
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
+
     /**
         Updated user
-    - PUT /user/{username}
+        - PUT /user/{username}
         - This can only be done by the logged in user.
         - API Key:
         - type: apiKey api_key 
         - name: api_key
         - parameter username: (path) name that need to be deleted 
         - parameter user: (body) Updated user object 
-    - returns: Future<Void, Never> 
+        - returns: AnyPublisher<Result<Void, Error>, Never> 
         */
-    open func updateUser(username: String, user: User) -> Future<Void, Never> {
+    open func updateUser(username: String, user: User) -> AnyPublisher<Result<Void, Error>, Never> {
+        // Creating final URL with query items and path
+        var components = URLComponents()
+        components.path = "/user/{username}"
+            .replacingOccurrences(of: "{username}", with: "\(username)")
+        
+        guard let url = components.url(relativeTo: transport.baseURL) else {
+            fatalError("URL is nil")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "PUT"
+        
+        // Setting body parameters 
+        request.httpBody = try? encoder.encode(user)
+        // Getting auth type 
+        let securityScheme: SecurityScheme = .none
+        return transport.send(request: request, securityScheme: securityScheme)
+            .map { response -> Result<Void, Error> in
+                if response.response?.statusCode == 200 {
+                    return .success(())
+                } else {
+                    return .failure(response.error ?? TransportError.unknown)
+                }
+            }
+            .eraseToAnyPublisher()
     }
 }
-////
 
-/**
--  
-    - API Key:
-    - type: apiKey api_key 
-    - name: api_key
-    - OAuth:
-    - type: oauth2
-    - name: petstore_auth
-- returns: RequestBuilder<Void> 
-    */
-open class func WithRequestBuilder() -> RequestBuilder<Void> {
-let localVariablePath = ""
-    let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-            let localVariableParameters: [String: Any]? = nil
 
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
+private class OpenISO8601DateFormatter: DateFormatter {
+    static let withoutSeconds: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.calendar = Calendar(identifier: .iso8601)
+        formatter.locale = Locale(identifier: "en_US_POSIX")
+        formatter.timeZone = TimeZone(secondsFromGMT: 0)
+        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZZZZZ"
+        return formatter
+    }()
 
-    let localVariableNillableHeaders: [String: Any?] = [
-        :
-    ]
-
-    let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-    let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-    return localVariableRequestBuilder.init(method: "", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+    private func setup() {
+        calendar = Calendar(identifier: .iso8601)
+        locale = Locale(identifier: "en_US_POSIX")
+        timeZone = TimeZone(secondsFromGMT: 0)
+        dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZZZZZ"
     }
 
-import Foundation
-#if canImport(AnyCodable)
-import AnyCodable
-#endif
+    override init() {
+        super.init()
+        setup()
+    }
 
-open class UserAPI {
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+        setup()
+    }
 
-    /**
-     Create user
-     
-     - parameter user: (body) Created user object 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func createUser(user: User, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
-        createUserWithRequestBuilder(user: user).execute(apiResponseQueue) { result in
-            switch result {
-            case .success:
-                completion((), nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
+    override func date(from string: String) -> Date? {
+        if let result = super.date(from: string) {
+            return result
         }
-    }
-
-    /**
-     Create user
-     - POST /user
-     - This can only be done by the logged in user.
-     - API Key:
-       - type: apiKey api_key 
-       - name: api_key
-     - parameter user: (body) Created user object 
-     - returns: RequestBuilder<Void> 
-     */
-    open class func createUserWithRequestBuilder(user: User) -> RequestBuilder<Void> {
-        let localVariablePath = "/user"
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: user)
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Creates list of users with given input array
-     
-     - parameter user: (body) List of user object 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func createUsersWithArrayInput(user: [User], apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
-        createUsersWithArrayInputWithRequestBuilder(user: user).execute(apiResponseQueue) { result in
-            switch result {
-            case .success:
-                completion((), nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Creates list of users with given input array
-     - POST /user/createWithArray
-     - API Key:
-       - type: apiKey api_key 
-       - name: api_key
-     - parameter user: (body) List of user object 
-     - returns: RequestBuilder<Void> 
-     */
-    open class func createUsersWithArrayInputWithRequestBuilder(user: [User]) -> RequestBuilder<Void> {
-        let localVariablePath = "/user/createWithArray"
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: user)
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Creates list of users with given input array
-     
-     - parameter user: (body) List of user object 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func createUsersWithListInput(user: [User], apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
-        createUsersWithListInputWithRequestBuilder(user: user).execute(apiResponseQueue) { result in
-            switch result {
-            case .success:
-                completion((), nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Creates list of users with given input array
-     - POST /user/createWithList
-     - API Key:
-       - type: apiKey api_key 
-       - name: api_key
-     - parameter user: (body) List of user object 
-     - returns: RequestBuilder<Void> 
-     */
-    open class func createUsersWithListInputWithRequestBuilder(user: [User]) -> RequestBuilder<Void> {
-        let localVariablePath = "/user/createWithList"
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: user)
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-        return localVariableRequestBuilder.init(method: "POST", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Delete user
-     
-     - parameter username: (path) The name that needs to be deleted 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func deleteUser(username: String, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
-        deleteUserWithRequestBuilder(username: username).execute(apiResponseQueue) { result in
-            switch result {
-            case .success:
-                completion((), nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Delete user
-     - DELETE /user/{username}
-     - This can only be done by the logged in user.
-     - API Key:
-       - type: apiKey api_key 
-       - name: api_key
-     - parameter username: (path) The name that needs to be deleted 
-     - returns: RequestBuilder<Void> 
-     */
-    open class func deleteUserWithRequestBuilder(username: String) -> RequestBuilder<Void> {
-        var localVariablePath = "/user/{username}"
-        let usernamePreEscape = "\(APIHelper.mapValueToPathItem(username))"
-        let usernamePostEscape = usernamePreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
-        localVariablePath = localVariablePath.replacingOccurrences(of: "{username}", with: usernamePostEscape, options: .literal, range: nil)
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters: [String: Any]? = nil
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-        return localVariableRequestBuilder.init(method: "DELETE", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Get user by user name
-     
-     - parameter username: (path) The name that needs to be fetched. Use user1 for testing. 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func getUserByName(username: String, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: User?, _ error: Error?) -> Void)) {
-        getUserByNameWithRequestBuilder(username: username).execute(apiResponseQueue) { result in
-            switch result {
-            case let .success(response):
-                completion(response.body, nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Get user by user name
-     - GET /user/{username}
-     - parameter username: (path) The name that needs to be fetched. Use user1 for testing. 
-     - returns: RequestBuilder<User> 
-     */
-    open class func getUserByNameWithRequestBuilder(username: String) -> RequestBuilder<User> {
-        var localVariablePath = "/user/{username}"
-        let usernamePreEscape = "\(APIHelper.mapValueToPathItem(username))"
-        let usernamePostEscape = usernamePreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
-        localVariablePath = localVariablePath.replacingOccurrences(of: "{username}", with: usernamePostEscape, options: .literal, range: nil)
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters: [String: Any]? = nil
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<User>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
-
-        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Logs user into the system
-     
-     - parameter username: (query) The user name for login 
-     - parameter password: (query) The password for login in clear text 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func loginUser(username: String, password: String, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: String?, _ error: Error?) -> Void)) {
-        loginUserWithRequestBuilder(username: username, password: password).execute(apiResponseQueue) { result in
-            switch result {
-            case let .success(response):
-                completion(response.body, nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Logs user into the system
-     - GET /user/login
-     - responseHeaders: [Set-Cookie(String), X-Rate-Limit(Int), X-Expires-After(Date)]
-     - parameter username: (query) The user name for login 
-     - parameter password: (query) The password for login in clear text 
-     - returns: RequestBuilder<String> 
-     */
-    open class func loginUserWithRequestBuilder(username: String, password: String) -> RequestBuilder<String> {
-        let localVariablePath = "/user/login"
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters: [String: Any]? = nil
-
-        var localVariableUrlComponents = URLComponents(string: localVariableURLString)
-        localVariableUrlComponents?.queryItems = APIHelper.mapValuesToQueryItems([
-            "username": username.encodeToJSON(),
-            "password": password.encodeToJSON(),
-        ])
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<String>.Type = OpenAPIClientAPI.requestBuilderFactory.getBuilder()
-
-        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Logs out current logged in user session
-     
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func logoutUser(apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
-        logoutUserWithRequestBuilder().execute(apiResponseQueue) { result in
-            switch result {
-            case .success:
-                completion((), nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Logs out current logged in user session
-     - GET /user/logout
-     - API Key:
-       - type: apiKey api_key 
-       - name: api_key
-     - returns: RequestBuilder<Void> 
-     */
-    open class func logoutUserWithRequestBuilder() -> RequestBuilder<Void> {
-        let localVariablePath = "/user/logout"
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters: [String: Any]? = nil
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-        return localVariableRequestBuilder.init(method: "GET", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
-    }
-
-    /**
-     Updated user
-     
-     - parameter username: (path) name that need to be deleted 
-     - parameter user: (body) Updated user object 
-     - parameter apiResponseQueue: The queue on which api response is dispatched.
-     - parameter completion: completion handler to receive the data and the error objects
-     */
-    open class func updateUser(username: String, user: User, apiResponseQueue: DispatchQueue = OpenAPIClientAPI.apiResponseQueue, completion: @escaping ((_ data: Void?, _ error: Error?) -> Void)) {
-        updateUserWithRequestBuilder(username: username, user: user).execute(apiResponseQueue) { result in
-            switch result {
-            case .success:
-                completion((), nil)
-            case let .failure(error):
-                completion(nil, error)
-            }
-        }
-    }
-
-    /**
-     Updated user
-     - PUT /user/{username}
-     - This can only be done by the logged in user.
-     - API Key:
-       - type: apiKey api_key 
-       - name: api_key
-     - parameter username: (path) name that need to be deleted 
-     - parameter user: (body) Updated user object 
-     - returns: RequestBuilder<Void> 
-     */
-    open class func updateUserWithRequestBuilder(username: String, user: User) -> RequestBuilder<Void> {
-        var localVariablePath = "/user/{username}"
-        let usernamePreEscape = "\(APIHelper.mapValueToPathItem(username))"
-        let usernamePostEscape = usernamePreEscape.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? ""
-        localVariablePath = localVariablePath.replacingOccurrences(of: "{username}", with: usernamePostEscape, options: .literal, range: nil)
-        let localVariableURLString = OpenAPIClientAPI.basePath + localVariablePath
-        let localVariableParameters = JSONEncodingHelper.encodingParameters(forEncodableObject: user)
-
-        let localVariableUrlComponents = URLComponents(string: localVariableURLString)
-
-        let localVariableNillableHeaders: [String: Any?] = [
-            :
-        ]
-
-        let localVariableHeaderParameters = APIHelper.rejectNilHeaders(localVariableNillableHeaders)
-
-        let localVariableRequestBuilder: RequestBuilder<Void>.Type = OpenAPIClientAPI.requestBuilderFactory.getNonDecodableBuilder()
-
-        return localVariableRequestBuilder.init(method: "PUT", URLString: (localVariableUrlComponents?.string ?? localVariableURLString), parameters: localVariableParameters, headers: localVariableHeaderParameters)
+        return OpenISO8601DateFormatter.withoutSeconds.date(from: string)
     }
 }
